feat(api): Implement login functionality
- Setup migrations script
- Update test script
- Write test to cover feature

[Finishes 161462334]
