+++ package.json
@@ -9,8 +9,9 @@
   "scripts": {
     "build": "rm -rf ./dist/* && babel ./server --out-dir ./dist --ignore ./server/test",
     "start": "npm run build && node ./dist",
+    "migrations":"babel-node ./server/config/migrations.js --env",
     "dev": "nodemon --exec babel-node ./server/index.js --env",
-    "test": "nyc --reporter=text --reporter=html mocha --recursive ./server/tests --exit --require babel-register"
+    "test": "npm run migrations && nyc --reporter=text --reporter=html mocha --recursive ./server/tests --exit --require babel-register"
   },
   "repository": {
     "type": "git",
